+++ CHANGELOG.md
@@ -1,3 +1,14 @@
-## 0.0.1
+#Changelog
+## [0.0.1] - 2024-08-10
+- package initiated
+- Basic functionality for overlapping circular images.
+## [0.0.2] - 2024-08-10
+- github link added to the homepage
+## [0.0.3] - 2024-08-10
+- user example code upload to the readme file
+## [0.0.4] - 2024-08-10
+- try to add some screenshot. 
+## [0.0.5] - 2024-08-10
+- example code added. 
 
 * TODO: Describe initial release.